Correcoes de layout

+++ caridade/src/main/java/br/com/caridade/repository/PastoralCaridadeRepository.java
@@ -8,6 +8,6 @@ import br.com.caridade.model.PastoralCaridade;
 
 public interface PastoralCaridadeRepository extends JpaRepository<PastoralCaridade, Long> {
 	List<PastoralCaridade> findByGtIsNotNullOrderByOrderGt();
-	List<PastoralCaridade> findBySemana(int semana);
+	List<PastoralCaridade> findBySemana(Long semana);
 
 }
